LayoutTests/inspector-protocol/dom/getAccessibilityPropertiesForNode.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/protocol-test.js"></script>
</head>
<body onLoad="runTest()">

<p>Checking Web Inspector protocol for the Accessibility Node Inspector.</p>

<div id="examples">

<script class="ex"></script>
<script class="ex" style="display:block;"></script>

<div class="ex"><div></div></div>
<div class="ex" role="presentation"><div></div></div>
<div class="ex"></div>
<div class="ex" aria-hidden="true"></div>
<span class="ex"></span>
<span class="ex" aria-hidden="true"></span>

<div class="ex" role="button" tabindex="0"></div>
<div class="ex" role="button" tabindex="0" aria-disabled="true">disabled</div>
<div class="ex" role="button" tabindex="0" aria-pressed="true">FIXME: Pressed is false. Expected true. http://webkit.org/b/129830</div>
<div class="ex" role="button" tabindex="0" aria-pressed="false">Not Pressed.</div>
<button class="ex"></button>
<input class="ex" type="button">

<!-- FIXME: Inputs missing axParentNodeId http://webkit.org/b/130181 -->
<input class="ex">
<input class="ex" required>
<input class="ex" aria-required="true" value="required">
<input class="ex" aria-invalid="spelling" value="invalid spelling">
<input class="ex" aria-invalid="foo" value="fake value will eval to true">
<input class="ex" readonly value="readonly">
<div class="ex" role="textbox" tabindex="0" aria-readonly="true">readonly</div>
<input class="ex" disabled value="disabled">
<div class="ex" role="textbox" tabindex="0" aria-disabled="true">disabled</div>

<input class="ex" type="checkbox">
<input class="ex" type="checkbox" checked>
<div class="ex" role="checkbox">unchecked (checked undefined evals to false on checkbox)</div>
<div class="ex" role="checkbox" aria-checked="true">checked</div>
<div class="ex" role="checkbox" aria-checked="mixed">mixed</div>
<div class="ex" role="checkbox" aria-checked="false">unchecked</div>

<input class="ex" type="radio">
<input class="ex" type="radio" checked>
<div class="ex" role="radio">unchecked (checked undefined evals to false on radio)</div>
<div class="ex" role="radio" aria-checked="true">checked</div>
<div class="ex" role="radio" aria-checked="mixed">unchecked [sic] mixed state will not be exposed on radios per ARIA spec.</div>
<div class="ex" role="radio" aria-checked="false">unchecked</div>

<input class="ex" aria-hidden="true">
<input class="ex" style="display:none;">

<!-- FIXME: Images missing axParentNodeId http://webkit.org/b/130181 -->
<img class="ex" src="./404.gif"><!-- img:not([alt]) ignored if it does not load.  -->
<img class="ex" src="data:image/gif;base64,R0lGODlhEAARAJECAOHh4UpKSgAAAAAAACH5BAEAAAIALAAAAAAQABEAAAIllB8Zx63b4otSUWcvyuz5D4biSD7AiZroWSXa5r7CJNOvra1RAQA7">
<img class="ex" src="data:image/gif;base64,R0lGODlhEAARAJECAOHh4UpKSgAAAAAAACH5BAEAAAIALAAAAAAQABEAAAIllB8Zx63b4otSUWcvyuz5D4biSD7AiZroWSXa5r7CJNOvra1RAQA7" alt="">
<img class="ex" src="data:image/gif;base64,R0lGODlhEAARAJECAOHh4UpKSgAAAAAAACH5BAEAAAIALAAAAAAQABEAAAIllB8Zx63b4otSUWcvyuz5D4biSD7AiZroWSXa5r7CJNOvra1RAQA7" alt="x">
<img class="ex" src="data:image/gif;base64,R0lGODlhEAARAJECAOHh4UpKSgAAAAAAACH5BAEAAAIALAAAAAAQABEAAAIllB8Zx63b4otSUWcvyuz5D4biSD7AiZroWSXa5r7CJNOvra1RAQA7" alt="x" aria-hidden="true">
<img class="ex" src="data:image/gif;base64,R0lGODlhEAARAJECAOHh4UpKSgAAAAAAACH5BAEAAAIALAAAAAAQABEAAAIllB8Zx63b4otSUWcvyuz5D4biSD7AiZroWSXa5r7CJNOvra1RAQA7" alt="x" style="display:none;">

<select multiple>
    <option class="ex" selected>FIXME: Selected is false. Expected true. http://webkit.org/b/129835</option>
    <option class="ex">not selected</option>
</select>
<div role="listbox">
    <div class="ex" role="option" aria-selected="true">selected</div>
    <div class="ex" role="option" aria-selected="false">not selected</div>
</div>
<div role="tablist">
    <div class="ex" role="tab" aria-selected="true">selected</div>
    <div class="ex" role="tab" aria-selected="false">not selected</div>
</div>

<ul role="tree">
    <li class="ex" role="treeitem" aria-expanded="true">expanded</li>
    <li class="ex" role="treeitem" aria-expanded="false">collapsed</li>
</ul>

</div>

<script type="text/javascript">

function $(selector) {
    return document.querySelectorAll(selector);
}
function cleanup() {
    // Hide the test element container to avoid irrelevant output diffs on subsequent updates.
    document.getElementById("examples").style.display = "none";
}

function test() {

    var examples = [];

    function loop() {
        if (examples.length) {
            InspectorTest.sendCommand("DOM.getOuterHTML", {"nodeId": examples[examples.length-1]}, onGotOuterHTML);
        } else {
            InspectorTest.log("\n");
            InspectorTest.sendCommand("Runtime.evaluate", {"expression": "cleanup()"}, function(){
                InspectorTest.completeTest();
            });
        }
    };

    function onGotDocument(response) {
        InspectorTest.checkForError(response);
        InspectorTest.sendCommand("DOM.querySelectorAll", {"nodeId": response.result.root.nodeId, "selector": ".ex"}, onQuerySelectorAll);
    }

    function onQuerySelectorAll(response) {
        InspectorTest.checkForError(response);
        examples = response.result.nodeIds;
        InspectorTest.log("Total elements to be tested: " + examples.length + ".");
        loop();
    }

    function onGotOuterHTML(response) {
        InspectorTest.checkForError(response);
        var outerHTML = response.result.outerHTML;
        outerHTML = outerHTML.replace(" class=\"ex\"", ""); // remove any duplicated, unnecessary class attributes
        outerHTML = outerHTML.replace(/;base64,.*?"/, "...\""); // suppress full data URIs in output
        InspectorTest.log("\n" + outerHTML);
        InspectorTest.sendCommand("DOM.getAccessibilityPropertiesForNode", {"nodeId": examples[examples.length-1]}, onGotAXProps);
    }

    function onGotAXProps(response) {
        InspectorTest.checkForError(response);
        for (var key in response.result.properties) {
            var value = response.result.properties[key];
            if (key === "nodeId")
                continue;
            if (key === "axParentNodeId")
                value = (value === 1) ? "1 (#document)" : "(suppressed)";
            InspectorTest.log("    " + key + ": " + value);
        }
        examples.pop();
        loop();
    }

    InspectorTest.sendCommand("DOM.getDocument", {}, onGotDocument);

}
</script>
</body>
</html>
