LayoutTests/inspector-protocol/dom/getAccessibilityPropertiesForNode-expected.txt

Checking Web Inspector protocol for the Accessibility Node Inspector.

Total elements to be tested: 51.

<li role="treeitem" aria-expanded="false">collapsed</li>
    exists: true
    label: 
    role: treeitem
    axParentNodeId: (suppressed)
    expanded: false

<li role="treeitem" aria-expanded="true">expanded</li>
    exists: true
    label: 
    role: treeitem
    axParentNodeId: (suppressed)
    expanded: true

<div role="tab" aria-selected="false">not selected</div>
    exists: true
    label: 
    role: tab
    axParentNodeId: (suppressed)

<div role="tab" aria-selected="true">selected</div>
    exists: true
    label: 
    role: tab
    axParentNodeId: (suppressed)
    selected: true

<div role="option" aria-selected="false">not selected</div>
    exists: true
    label: 
    role: option
    axParentNodeId: (suppressed)

<div role="option" aria-selected="true">selected</div>
    exists: true
    label: 
    role: option
    axParentNodeId: (suppressed)
    selected: true

<option>not selected</option>
    exists: false
    label: 
    role: 

<option selected="">FIXME: Selected is false. Expected true. http://webkit.org/b/129835</option>
    exists: false
    label: 
    role: 

<img src="data:image/gif..." alt="x" style="display:none;">
    exists: false
    label: 
    role: 

<img src="data:image/gif..." alt="x" aria-hidden="true">
    exists: true
    label: 
    role: img
    ignored: true
    ignoredByDefault: true
    hidden: true

<img src="data:image/gif..." alt="x">
    exists: true
    label: 
    role: img

<img src="data:image/gif..." alt="">
    exists: true
    label: 
    role: img
    ignored: true

<img src="data:image/gif...">
    exists: true
    label: 
    role: img

<img src="./404.gif">
    exists: true
    label: 
    role: img
    ignored: true

<input style="display:none;">
    exists: false
    label: 
    role: 

<input aria-hidden="true">
    exists: true
    label: 
    role: 
    focused: false
    ignored: true
    ignoredByDefault: true
    hidden: true
    required: false

<div role="radio" aria-checked="false">unchecked</div>
    exists: true
    label: 
    role: radio
    axParentNodeId: 1 (#document)
    checked: false
    required: false

<div role="radio" aria-checked="mixed">unchecked [sic] mixed state will not be exposed on radios per ARIA spec.</div>
    exists: true
    label: 
    role: radio
    axParentNodeId: 1 (#document)
    checked: false
    required: false

<div role="radio" aria-checked="true">checked</div>
    exists: true
    label: 
    role: radio
    axParentNodeId: 1 (#document)
    checked: true
    required: false

<div role="radio">unchecked (checked undefined evals to false on radio)</div>
    exists: true
    label: 
    role: radio
    axParentNodeId: 1 (#document)
    checked: false
    required: false

<input type="radio" checked="">
    exists: true
    label: 
    role: radio
    checked: true
    focused: false
    required: false

<input type="radio">
    exists: true
    label: 
    role: radio
    checked: false
    focused: false
    required: false

<div role="checkbox" aria-checked="false">unchecked</div>
    exists: true
    label: 
    role: checkbox
    axParentNodeId: 1 (#document)
    checked: false
    required: false

<div role="checkbox" aria-checked="mixed">mixed</div>
    exists: true
    label: 
    role: checkbox
    axParentNodeId: 1 (#document)
    checked: mixed
    required: false

<div role="checkbox" aria-checked="true">checked</div>
    exists: true
    label: 
    role: checkbox
    axParentNodeId: 1 (#document)
    checked: true
    required: false

<div role="checkbox">unchecked (checked undefined evals to false on checkbox)</div>
    exists: true
    label: 
    role: checkbox
    axParentNodeId: 1 (#document)
    checked: false
    required: false

<input type="checkbox" checked="">
    exists: true
    label: 
    role: checkbox
    checked: true
    focused: false
    required: false

<input type="checkbox">
    exists: true
    label: 
    role: checkbox
    checked: false
    focused: false
    required: false

<div role="textbox" tabindex="0" aria-disabled="true">disabled</div>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    disabled: true
    focused: false
    readonly: true
    required: false

<input disabled="" value="disabled">
    exists: true
    label: 
    role: 
    disabled: true
    required: false

<div role="textbox" tabindex="0" aria-readonly="true">readonly</div>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    focused: false
    readonly: true
    required: false

<input readonly="" value="readonly">
    exists: true
    label: 
    role: 
    focused: false
    readonly: true
    required: false

<input aria-invalid="foo" value="fake value will eval to true">
    exists: true
    label: 
    role: 
    focused: false
    invalid: true
    required: false

<input aria-invalid="spelling" value="invalid spelling">
    exists: true
    label: 
    role: 
    focused: false
    invalid: spelling
    required: false

<input aria-required="true" value="required">
    exists: true
    label: 
    role: 
    focused: false
    required: true

<input required="">
    exists: true
    label: 
    role: 
    focused: false
    required: true

<input>
    exists: true
    label: 
    role: 
    focused: false
    required: false

<input type="button">
    exists: true
    label: 
    role: button
    focused: false

<button></button>
    exists: true
    label: 
    role: button
    focused: false

<div role="button" tabindex="0" aria-pressed="false">Not Pressed.</div>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    focused: false
    pressed: false
    required: false

<div role="button" tabindex="0" aria-pressed="true">FIXME: Pressed is false. Expected true. http://webkit.org/b/129830</div>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    focused: false
    pressed: false
    required: false

<div role="button" tabindex="0" aria-disabled="true">disabled</div>
    exists: true
    label: 
    role: button
    axParentNodeId: 1 (#document)
    disabled: true
    focused: false

<div role="button" tabindex="0"></div>
    exists: true
    label: 
    role: button
    axParentNodeId: 1 (#document)
    focused: false

<span aria-hidden="true"></span>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    ignored: true
    ignoredByDefault: true
    hidden: true

<span></span>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    ignored: true
    ignoredByDefault: true

<div aria-hidden="true"></div>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    ignored: true
    ignoredByDefault: true
    hidden: true

<div></div>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    ignored: true

<div role="presentation"><div></div></div>
    exists: true
    label: 
    role: presentation
    axParentNodeId: 1 (#document)
    ignored: true

<div><div></div></div>
    exists: true
    label: 
    role: 
    axParentNodeId: 1 (#document)
    ignored: true

<script style="display:block;"></script>
    exists: true
    label: 
    role: group
    axParentNodeId: 1 (#document)
    ignored: true

<script></script>
    exists: false
    label: 
    role: 
